hardware/requirements.txt
websockets>=11.0.3
python-escpos>=3.0.0
pyserial>=3.5
